Fixed #65
Resolved the issue with the hidden field. The jQuery wasn't implemented
that set the checkbox value to 'true' when it was checked.  This is how
the framework gets the value of the checkbox field. This jQuery code not
implemented caused the value to always be false when it was checked or
not checked